+++ backend/src/web/__init__.py
@@ -10,3 +10,15 @@ from .reverse_image_search_crawler import (
 from .dropbox_drive_sync import DropboxDriveSync as DropboxDriveSync
 from .google_drive_sync import GoogleDriveSync as GoogleDriveSync
 from .one_drive_sync import OneDriveSync as OneDriveSync
+
+__all__ = [
+    "ImageCrawler",
+    "DanbooruCrawler",
+    "GelbooruCrawler",
+    "SankakuCrawler",
+    "WebRequestsLogic",
+    "ReverseImageSearchCrawler",
+    "DropboxDriveSync",
+    "GoogleDriveSync",
+    "OneDriveSync",
+]